+++ src/main/java/com/guxingke/intellij/plugin/Const.java
@@ -1,8 +1,8 @@
 package com.guxingke.intellij.plugin;
 
+import java.util.Set;
+
 public interface Const {
 
-  String CLS_JAVA_LANG_STRING = "java.lang.String";
-  String CLS_JAVA_UTIL_COLLECTION = "java.util.Collection";
-  String CLS_JAVA_UTIL_STREAM_STREAM = "java.util.stream.Stream";
+  Set<String> PREDEFINED_VARS = Set.of("expr", "exprClass", "componentClass", "END");
 }