+++ src/main/java/com/guxingke/intellij/plugin/postfix/template/ScriptPostfixTemplate.java
@@ -1,11 +1,19 @@
 package com.guxingke.intellij.plugin.postfix.template;
 
+import com.guxingke.intellij.plugin.Const;
 import com.guxingke.intellij.plugin.postfix.template.conf.TemplateDefinition;
+import com.guxingke.intellij.plugin.util.VariableUtils;
+import com.guxingke.intellij.plugin.postfix.Var;
 import com.guxingke.intellij.plugin.util.PsiExpressionUtils;
 import com.intellij.codeInsight.template.Template;
 import com.intellij.codeInsight.template.TemplateManager;
+import com.intellij.codeInsight.template.impl.SelectionNode;
 import com.intellij.codeInsight.template.impl.TextExpression;
 import com.intellij.codeInsight.template.postfix.templates.PostfixTemplateProvider;
+import com.intellij.notification.Notification;
+import com.intellij.notification.NotificationGroupManager;
+import com.intellij.notification.NotificationType;
+import com.intellij.notification.Notifications;
 import com.intellij.openapi.util.Condition;
 import com.intellij.psi.JavaPsiFacade;
 import com.intellij.psi.PsiArrayType;
@@ -14,6 +22,8 @@ import com.intellij.psi.PsiClassType;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiExpression;
 import com.intellij.psi.PsiPrimitiveType;
+import java.util.Comparator;
+import java.util.stream.Collectors;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -88,12 +98,53 @@ public class ScriptPostfixTemplate extends BasePostfixTemplate {
                                    facade
     );
     var item = TemplateFactory.match(ctx, definition.getTemplates());
+    assert item != null;
 
-    var tpl = manager.createTemplate(getId(), "", item.getTemplate());
+    var ts = item.getTemplate();
+    // parse variables
+    var vars = VariableUtils.parseVariables(ts)
+        .stream()
+        .filter(it -> !Const.PREDEFINED_VARS.contains(it.getName()))
+        .sorted(Comparator.comparingInt(Var::getNo))
+        .collect(Collectors.toList());
+
+    var nts = VariableUtils.removeVariableValues(ts, vars);
+
+    var tpl = manager.createTemplate(getId(), "", nts);
     tpl.addVariable("expr", new TextExpression(e.getText()), false);
     tpl.addVariable("exprClass", new TextExpression(ctx.getClazzName()), false);
     if (ctx.getComponentClass() != null) {
       tpl.addVariable("componentClass", new TextExpression(ctx.getComponentClassName()), false);
+    }
+    // others values
+    for (Var var : vars) {
+      try {
+        tpl.addVariable(var.getName(),
+                        var.getExpression(),
+                        var.getDefaultValueExpression(),
+                        var.isAlwaysStopAt(),
+                        var.skipOnStart()
+        );
+
+      } catch (Exception exception) {
+        tpl.addVariable(var.getName(),
+                        new SelectionNode(),
+                        new SelectionNode(),
+                        var.isAlwaysStopAt(),
+                        var.skipOnStart()
+        );
+
+        var notificationGroup = NotificationGroupManager.getInstance().getNotificationGroup("The One Toolbox");
+        Notification notification = notificationGroup.createNotification("Error in postfix template",
+                                                                         "Your " + tpl.getKey()
+                                                                             + " template contains an error in variable '"
+                                                                             + var.getName() + "'. Please fix it.",
+                                                                         NotificationType.ERROR
+        );
+
+        Notifications.Bus.notify(notification, project);
+
+      }
     }
 
     return tpl;