+++ src/test/java/com/guxingke/intellij/plugin/util/VariableUtilsTest.java
@@ -1,0 +1,13 @@
+package com.guxingke.intellij.plugin.util;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class VariableUtilsTest {
+
+  @Test
+  public void parseVariableNames() {
+    var vals = VariableUtils.parseVariableNames("$expr$.alsdkfjla.$kesj$a.sdfalsd.eo$valk$$END$");
+    Assert.assertTrue(vals.size() > 0);
+  }
+}